Improvements to cpp intellisense
